move era5 land to era5.py
